server_api/requirements.txt
attrs==21.4.0
beautifulsoup4==4.11.1
blis==0.7.8
catalogue==2.0.8
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
confection==0.0.2
configparser==5.3.0
conllu==4.5.2
cymem==2.0.6
docopt==0.6.2
fastjsonschema==2.15.3
Flask==2.2.2
idna==3.3
importlib-metadata==5.0.0
importlib-resources==5.7.1
interchange==2021.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter-core==4.10.0
langcodes==3.3.0
MarkupSafe==2.1.1
monotonic==1.6
murmurhash==1.0.8
nbformat==5.4.0
nltk==3.7
nmslib==2.1.1
num2words==0.5.12
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pansi==2020.7.3
pathy==0.6.2
preshed==3.0.7
psutil==5.9.2
py2neo==2021.2.3
pybind11==2.6.1
pycryptodome==3.15.0
pydantic==1.9.2
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
pysbd==0.3.4
python-aiml==0.9.3
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
regex==2022.8.17
requests==2.27.1
scikit-learn==1.1.2
scipy==1.9.1
scispacy==0.5.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
thinc==8.1.2
threadpoolctl==3.1.0
tqdm==4.64.1
traitlets==5.2.1.post0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.9
wasabi==0.10.1
Werkzeug==2.2.2
zipp==3.8.0
